Downloads/chromadb-llm-performance-comparison-rag/requirements.txt
langchain
langchain-community
chromadb
openai
anthropic
llama-index
streamlit
pandas
numpy
sqlite-utils
python-dotenv
pypdf
